separated Excel rendering from mailing (single responsibility, test isolation)
